[turboshaft] JSStackCheck can deopt and throw (unlike WasmStackCheck)
This CL splits JS and Wasm stack checks into 2 different operations,
since JS stack checks can throw and deopt (unlike Wasm stack checks).

StackCheckOp was so far rarely used in the JS pipeline (since Turbofan
lowers JSStackChecks in JSGenericLowering, before reaching
Turboshaft). However, with Maglev as a frontend, we can get
StackCheckOp in the JS pipeline, and we need to make sure that it has
a frame state (to be able to lazy deopt) and a valid Context (to
throw).

This also fixes a bug with --trace-deopt on functions with no deopt
point but that could still deopt because of protector invalidation
(still with the Maglev frontend only): their DeoptimizationData was so
far empty, which lead to a DCHECK failure in --trace-deopt. Now, the
DeoptimizationData of a JS function can never be empty, since at least
the initial JSStackCheckOp can deopt.

Bug: chromium:42204525
Change-Id: I784f104f47178f44401ca7215bc7046f7ca261ea
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5621891
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94438}