[turboshaft] Maglev-to-ts: reorganize mjsunit tests
Having (almost) all tests in a single file is not great to investigate
failures (for instance, if one of the tests in the old
turboshaft-maglev-frontend.js fails, running with --trace-turbo-graph
can print hundred of thousands of lines before reaching the
interesting function) and means that we have to specify more flags
than necessary.

This CL thus splits turboshaft-maglev-frontend.js into multiple
smaller tests (the content of the tests remained the same though).

Bug: chromium:42204525
Change-Id: Ic9ec58ab0d43ac4c693b9ca9906d9b58c6e221f6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5621246
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94440}