[inspector] Change 'description' for remote Error objects
Until now the 'description' for Error objects contains different
content depending on whether the Error originates from V8 itself or
is an embedder specific Error (e.g. DOMException). For "native" Errors
we would simply use Error#stack. This is not always ideal, since
Error#stack is cemented at the time the Error object is created.

There is a need to support Error objects where 1) the "message"
property is populated after the fact, and 2) custom Error subclasses
that don't populate the "name" property.

This CL changes the 'description' to the following pattern:
<name property> or <constructor name> ": " <message property>
  <stack property without first line>

We use either the value of the "name" property or the constructor name
if "name" is the default "Error". This is to both stay consistent
with Error.toString but also support the common case of Error
sub-classes that don't set the "name" property.

The only iffy bit is that we cut the Error#stack of via a string
pattern, so if the "message" property contains "\n    at " then we
would include parts of the "message" twice.

Bug: 344629939
Change-Id: I9bdb8ffbb37a9d01d90eeec4e3bff4b9fae9061a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5633534
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94505}