[liftoff] Use ptrsized comparison to detect OOB on table64
This fixes OOB detection on 64-bit platforms for table64. The stored
table size is always a 32-bit value, but if the dynamic index is a
64-bit value we need to do a 64-bit comparison.

It does not matter if we zero-extend or sign-extend the actual table
size, because it is bound by kV8MaxWasmTableSize (currently 10M). Thus
we can treat it as a 32-bit signed immediate (if the table size is
constant) in code generation.

R=mliedtke@chromium.org
CC=evih@chromium.org

Bug: 338024338
Change-Id: I76c9c33021fc1c0e71acda0b5158185dddcdc973
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5633596
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94460}