[wasm64] Speed-up trap-handling based bounds checks
Mitigate perf regression reported in v8/SciMark2.1 after enabling a new
bounds checking mechanism based on trap handling.
On Wasm64 we cannot totally rely on guard pages but we also need to trap
when `index >= memory_guards_size`. The regressing implementation was
implementing this trap with a branch that modified `index` to force an
access into a guard page, but the generated IR and code was too
verbose in large functions that have a large number of loads and stores
(like FFT in SciMark2.1). Instead, it is faster to directly trap when
`index >= memory_size`.

Bug: 345936778
Change-Id: I226ffc07c483e3cf2f543efc798f74c352cd0264
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5634283
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#94475}