[heap] Prepare ArrayBufferSweeper for sticky markbits
When detaching an ArrayBufferExternsion, we check the age of the
JSArrayBuffer to know which list of ArrayBuffers should contains the
extension. With sticky markbits, if detaching an extension during full
GC marking, the age check cannot be relied on. All objects are unmarked
and treated as old during marking, yet the extension may reside in the
young list.

Instead we need to find the list containing the extension without asking
the JSArrayBuffer. This CL achieves that by storing the age in the
extension, updating it when promoting the extension, and querying it
when detaching.

See [1] for further discussion of the issue and alternative solutions.

[1] https://docs.google.com/document/d/1HlzSGF36y-7njo6UIe3eOJi6CpQdoXskjVR8JGO8N0I/edit?usp=sharing

Bug: 346918506
Change-Id: I2afadd90127e109dd0a61cd61d4cd75e09cea2b4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5626413
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94484}