[parser] Allocate class function ids before their inner functions
static elements and instance member functions wrap their respective
class elements. Currently they are allocated after the class is fully
parsed; which is also when their function id is allocated. This CL
moves function ID allocation to when the scope is created as well,
that is, right before the first AstNode that will be moved into it is
parsed. That ensures that the function IDs for the respective methods
are lower than their inner functions' IDs.

Change-Id: Icecbaaa7274af3db7e3a3c1b177d09ab91c81569
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5633537
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94495}