[cppgc] Delay cross-thread root marking
Cross-thread roots (CrossThreadPersistent and
WeakCrossThreadPersistent) marking and weakness clearing requires
obtaining a lock from the GC during the final atomic pause. The same
lock is used for creating these references. This way we ensure that
the GC doesn't reclaim objects that are held by other threads. It also
ensures that converting a weak into a strong persistent is consistent
wrt to GC running on a different thread.

In order to minimize the impact of this lock we already delayed the
acquisition in GC till we were done marking all CppHeap objects.
However, there may be a lot of JS objects left which can cause quite a
large window where this lock is taken.

This CL only acquires the lock at the very end of marking when
everything else is marked. There will usually not be many objects left
to be marked from this root set.

In addition, we clear cross-thread weak roots immediately durng weak
roots processing which allows us to further shrink down the time
window where the GC holds the lock.

Bug: 347686125, 345117891
Change-Id: I6e154afdfb6cf7d0b24b9f6a5a237140dfc90ed0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5630528
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94532}