[turboshaft] Fix wrong-removing of non-empty loops
Turboshaft's loop unrolling has some logic to remove loops with 0
iterations, and it was broken.

The idea is that is there is a loop whose termination condition is
something like "stop when i <= 0" where `i` is a loop phi and `i`'s
initial value is less than or equal to 0, then this loop has 0
iterations. EXCEPT that this is only true if `i` is a loop phi FOR
THIS LOOP. And, in the repro, the `i` comes from a previous loop, so
what matters is not its initial value but rather its "final" value
(which is not known unless we know how many iterations this previous
has, but regardless, the current loop doesn't have 0 iterations!).

Fixed: chromium:347905990
Bug: chromium:42202729
Change-Id: I8cb80d9e265591f56ce1a32219a0830148216dde
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5640927
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94543}