[liftoff] Allow constant memory and table indexes
The {PopIndexToVarState} and {IndexToVarStateSaturating} methods
currently materialized 32-bit constants into a register on 64-bit
machines. As this is the most common case (memory32 and table32 are the
default), this CL allows constants to stay constants if they are 31-bit
values (i.e. the MSB is not set).
This will generate shorter code for table and memory operations with
constant indexes, including indirect calls.

R=evih@chromium.org

Bug: 338024339, 42204510
Change-Id: I7508d4b27da9a2533439fdb1727669e8741a3f7e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5640160
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94563}