[fuzzer] Init expr fuzzer: Use liftoff only
This is a follow-up to https://crrev.com/c/5630312 trying to prevent the
fuzzer running into timeouts. Fortunately, inlining is not responsible
for the timeouts (as there also shouldn't be any inlining candidates).

As a next step, disable Turbofan for the init expr fuzzer. The
compilation of these function is anyways only needed to verify the
correctness of the result produced by the init expression (which is
interpreted), so it doesn't matter which compiler tier is used to
produce the expected result.

Change-Id: I82a47c7cfeeac8aabbb7ce7a8081fcc53a6deaee
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5639900
Commit-Queue: Eva Herencsárová <evih@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94558}