[wasm] Split enabled and detected features
This makes it a lot easier to find all the places to clean up when
removing a feature flag. They just won't compile anymore, and we have to
decide whether to add something to the detected set instead.

This CL already forces us to remove a few DCHECKs and flag scopes for
always-enabled features, which is a good thing.

R=mliedtke@chromium.org

Change-Id: If2f57df52da1f91c5d2955e05878c8bcec9bb286
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5631031
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94553}