[deoptimizer][wasm] Allocate feedback vector during deopt if missing
In cases of multiple instantiations of the same module it is possible
that a function is deopted for which the feedback vector has not been
allocated yet.
The deoptimizer needs to allocate it, however during deoptimization a
heap allocation is not possible as the stack is in a non-gc-safe state.
Therefore, the deoptimizer continues at an additional piece of liftoff
code that calls a builtin (with safetpoint information) to allocate the
feedback vector.

Note: This change only handles this for the innermost stack frame.
Any other inlined frames may also miss feedback vectors which will
be added in a follow-up change.

Bug: 42204618
Change-Id: I6c67fb3b05061b2e041978dc541674c0191c7fb3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5643118
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94587}