[api] Support preprocessing of exceptions thrown from Api callbacks
This is a first bulk of changes adding an experimental Api for
preprocessing exceptions thrown from Api callbacks. This a part of
the project aimed at reducing the overhead of blink::ExceptionState
objects on Chromium side.

This CL adds some plumbing allowing the new callback to get details
about the last crossing of JS -> Api boundary which can be used
for extending exception message if necessary.

Bug: 328104148
Change-Id: I57806f4511602e7b2092bcb683149c62073a4cc4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5526293
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94588}