[deoptimizer][wasm] Delete deoptimizer from liftoff via builtin
Instead of deleting the deoptimizer during the deoptimization in the
Generate_DeoptimizationEntry builtin which is platform-specific,
re-use the extra builtin call from liftoff directly after the deopt to
perform the C++ object destruction.

This way the Deoptimizer is also around long enough to be used within
that builtin (e.g. to read the number of liftoff frames for which a
feedback vector may need to be allocated, see
https://crrev.com/c/5643118 for the allocation for a single frame which
needs to be extended for multiple frames.

In the future, this builtin could also be re-used for lazy wasm object
materialization (which is a feature that JS supports). However, this
would require much more infrastructure for wasm deopts and is out of
scope for now.

Bug: 42204618
Change-Id: I7edbc23eb8b9bac40c0f8bfe4b1fc424c468d568
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5644892
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94596}