[cppgc] Sweep based on live bytes on pages
Gardeners: This may cause a slight regression for a feature that
landed behind field trial (V8GCOptimizeSweepForMutator) as a
progression. The overall situation is still a net win once launched.

Not sweeping on allocation can cause memory regressions in case where
there is no progress via tasks as we are effectively getting new
pages. We also need sweeping to return empty pages due to having finalizers.

This CL brings back sweeping on allocation instead of bailing
out depending on the limit. At the same time we introduce per-page
live byte counts which are populated during marking. The sweeper can
then sort pages accordingly and start with pages that have the most
free bytes. This results in lower time spent in sweeping on allocation
as the chance is high to get a large block or almost empty page here.

In future, we can adjust concurrent sweeper to be page agnostic which
should improve this even further.

Change-Id: I8c6d4cd037c8fb1b521d1a43606d673fe0d370a1
Bug: 333981063
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5645442
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94597}