[api] Cleanup usages of v8::ReturnValue<void>::Set[NonEmpty](..)
The Setter/Definer interceptor callbacks used to call this methods in
order to indicate that the operation was intercepted, however after
recent change of interceptor callbacks signatures this is done
explicitly by returning v8::Intercepted value.

Also, old Api tests were often returning the property's value that
was intended to be set or defined. Moreover, V8 was interpreting
this value as a boolean result indicating whether the operation
succeeded or not which is incorrect.

On the other hand, according to JavaScript spec [[Set]] and
[[DefineOwnProperty]] operations' result is a boolean value, so the
callbacks need a way to indicate that the operation was not successful.
This result is observable via Reflect.set() or
Reflect.defineProperty().

This CL
 - deprecates v8::ReturnValue<void>::Set[NonEmpty](..),
 - makes 'true_value' the default return value for Deleter callbacks,
   so it's not necessary to call Set(true) to indicate success,
 - allows using Set(bool) and SetFalse() to indicate the result of
   [[Set]] or [[DefineOwnProperty]] operations,
 - updates various Api tests misusing v8::ReturnValue<void>::Set(..)
   methods,
 - introduce enum InterceptorResult so that Maybe<InterceptorResult>
   is able to represent all four possible outcomes of Setter/Definer/
   Deleter interceptor callbacks:
    - not intercepted,
    - intercepted + thrown exception,
    - intercepted + success/failure.

Bug: 348660658
Change-Id: I6227173a1e6045bfca44fb81a26e800ae4943a67
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5647894
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94631}