Adds api to String to internalize a string
Some code in blink is putting a v8::String in a map and using
the string as a key. To ensure strings that are equal have the
same hash-code, the code needs to use internalized strings. This
exposes the API to get the internalized string.

Blink patch is here:
https://chromium-review.googlesource.com/c/chromium/src/+/5647781

Bug: 347609263
Change-Id: I92c8babd788c5653888b206f57139c8db41062cb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5651775
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94641}