Don't DCHECK when setting a WASM breakpoint on a non-zero line number
This stops a crash when passing a non-zero line number via CDP when
setting WASM breakpoints in DCHECK enabled builds. Instead we fail
to resolve the breakpoint location.

We cannot provide a more meaningful error message via CDP as the
places that check the line number and where we have access to
the CDP response object are a couple stack frames apart.

R=bmeurer@chromium.org

Fixed: 349248170
Change-Id: I0f575113fd1bbe590cd8cc472c2c9522ffe3e1d1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5658170
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94647}