Add HeapProfiler::IsTakingSnapshot to public API
The function HeapProfiler::CopyNameForHeapSnapshot only works while V8
is taking a heap snapshot, but there is currently no way for the
embedder to tell whether a snapshot is in progress, so nobody can safely
use CopyNameForHeapSnapshot. To remedy the problem, this change adds a
function to check whether the heap profiler is taking a snapshot.

Bug: 343341610
Change-Id: I1e0c40a27b192d241649ac2a2dcefa53d823b971
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5656083
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94655}