[arm64][simulator] Fix debugger not recognizing special x registers
Debugging experience:
  # Debugger hit 0: DebugBreak
  #                 LOCATION: DeoptimizationEntry_Eager
  0x00007faff76002fc  91044030            add x16, x1, #0x110 (272)
  sim> print x1
   x1: 0x0000556ad5174a10
  sim> print x16
  x16 unrecognized
  sim>

It'd be nice if the disassembler and the simulator had a similar
understanding of register names. This CL fixes this issue for the x
registers with special names.

This change does not change the printed name, i.e. the new output is:
  0x00007f9cf76002fc  91044030            add x16, x1, #0x110 (272)
  sim> print x16
   ip0: 0x00007f9c402baf70

Change-Id: I68cae77bd7f59af9ee9e86fb244c4ae5e41035bf
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5659605
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94702}