[heap] Fix JSGlobalProxy reinitialization
Upon reinitialization we write into an object that already has been
published. As such the object is visible to the marker. It is
important that we don't overwrite CppHeapPointerHandle with undefined
values as that's not a valid sentinel handle. Instead, the marker
would interpret the undefined value as handle and likely crash.

In the unlikely case that the handle already exists, we violate the
invariant that handles are unique which is necessary for correctness
in the compaction of the CppHeapPointerTable.

Bug: 349797714, 343889872
Change-Id: I34ebfd7060b618552e72d3ec08130cdb5b4e62f2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5663185
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94708}