[top-level-await] Add test case for transition bug; beef up --trace-module-status
The actual fix was in https://crrev.com/c/5661645

When transitioning components to the EVALUATED state, modules that are
in the middle of async evaluation should transition to EVALUATING-ASYNC,
while modules not async evaluating should transition to EVALUATED. There
was a bug that both incorrectly transitioned entire components to
EVALUATING-ASYNC and did not set the correct [[CycleRoot]] on modules in
the components, resulting in dereferencing the hole value.

As a drive-by, beef up --trace-module-status, which was used to debug
this.

Fixed: 349549076
Change-Id: Ic83e6e94a642c624026b603593183e003d85d2d6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5664994
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94715}