Fix shared-string cctest failures with v8_enable_builtins_profiling
When an Isolate is deinitializing, if the build included profiling of
builtins, then the Isolate attempts to read some profiling data from the
heap. That data on the heap also includes some strings representing the
names of the builtins. Those strings are stored in shared space. If the
Isolate detaches itself from the shared space before reading those
strings, then there is a race condition where another Isolate might run
garbage collection on the shared heap and destroy those strings. To fix
the problem, this CL moves the problematic step earlier during disposal
of the Isolate.

Bug: 344654361
Change-Id: I82aeda7ac415110fdb28e50400249b31668511c6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5664995
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#94721}