[wasm] Fix invalid use of the generic wasm-to-js wrapper
When creating a WasmFuncRef, we assumed that the callee was a regular JS
function and used the generic wasm-to-js wrapper if the signature
matched. However some import call kinds are not supported by the generic
wrapper such as constructors. Recompute the import call kind and use the
optimized wasm-to-js wrapper if needed.

R=ahaas@chromium.org

Fixed: 348963663
Change-Id: Ibd5b71af4b5ce14f73b63db1eaf4e7b9a02aab76
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5664935
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94698}