[sandbox] Check functionData type in generic JSToJSWrapper
In crrev.com/c/5591701 we added a SBXCHECK into the generic
JSToWasmWrapper to make sure that the 'data' field loaded from the SFI
contains a WasmFunctionData object (which in this case is necessary as
the trusted pointer field uses the wildcard kUnknownIndirectPointerTag).
This CL does the same for the JSToJSWrapper.

Bug: 343749048
Change-Id: I9c11c9887e939e01d4cf3e35fd8ea53fe0c63fe8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5664927
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94768}