Clean up constructors of Managed
The `Managed<T>` class provided four constructors, for constructing from
- arguments to pass to the constructor of `T`,
- a raw pointer to `T`,
- a unique_ptr to `T`,
- a shared_ptr to `T`.

As we internally always store a `T` anyway it is cleaner to accept just
that when constructing a `Managed`. For the caller it's easy enough to
convert whatever they have or whatever whey are allocating to a
`shared_ptr`.

R=leszeks@chromium.org

Bug: 349563054
Change-Id: I598d6c6887ea148106a09a0b0360effc83365dca
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5670365
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94746}