[wasm] Store link to NativeModule in TrustedManaged
We need to make sure that the NativeModule stays alive as long as its
code is being executed. We thus store a TrustedManaged to the
NativeModule on the WasmTrustedInstanceData. The latter is kept alive by
a reference on the stack by every frame which still executes for of that
instance.

R=jkummerow@chromium.org
CC=saelo@chromium.org

Fixed: 349563054
Change-Id: I557337faa4bee001c375176247606164c2614b87
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5666885
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94774}