[wasm][sandbox] Harden data segment initialization
During construction of the WasmTrustedInstanceData we only want to read
the link from WasmModuleObject to the NativeModule once to ensure
consistency throughout the construction of the trusted data. The link is
then stored in a TrustedManaged.

This CL removes another path where we read the link again for
initializing the data segments.

R=jkummerow@chromium.org

Bug: 336507783
Change-Id: Idbcbc859646a15138e7e22c876f2f88690f2d7a2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5666887
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94777}