[sandbox] Make SBXCHECK in JSHeapBroker::TryGetOrCreateData more precise
Previously, we were only checking that if IsX was true for the source
HeapObject, then IsX should also be true for the corresponding
ObjectData that we create from it. However, this is too coarse, as for
example in the generic HeapObject case, IsHeapObject will be true even
if the serialized object reports a different instance type than the
original object. In particular, it can happen that we serialize a
non-Map object, which then however reports MAP_TYPE as its instace type
as its map pointer points to itself (see the special case in
HeapObjectData::GetMapInstanceType()). The new check is now more
precise: we check that the InstanceType read from the heap and the one
reported by the serialized object are exactly equal.

Bug: 349641090
Change-Id: Ifdbf2324fb1a419ec25eb684c9601500bfe1035b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5674421
Reviewed-by: Stephen Röttger <sroettger@google.com>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94833}