Reland "[common] Improve per-thread assert scopes"
This is a reland of commit 42c92b8634ed39a9167eae5d090c51094d72bb8f

This reland fixes gcmole by adding template parameters to
PerThreadAssertScopeEmpty, which keeps debug-only scopes as distinct
types from each other in release builds. This is needed for the
DisableGCMole scope to be recognised as distinct from other scopes.

Also, since I'm relanding anyway, this simplifies the clearing logic
for PerThreadAssertData to use an always-on bit, rather than a cleared
bit, so that clearing is just zeroing the value rather than setting a
bit on the value (and checking for cleared is checking for zero, which
is a tiny bit cheaper to encode than testing for a bit).

Original change's description:
> [common] Improve per-thread assert scopes
>
> Three enhancements for per-thread assert scopes:
>
>   * Assert scopes are now variadic in the PerThreadAssertType they
>     enable/disable. This means that combination scopes no longer need to
>     exist, and are replaced by single variadic scopes that set multiple
>     bits at the same time. To make this simpler, the per-thread data
>     becomes an EnumSet instead of multiple BitFields.
>
>   * The std::optional in PerThreadAssertScope is gone, replaced by a
>     marker value in the PerThreadAssertType enum. This makes scopes
>     4 bytes instead of 8 (with padding).
>
>   * PerThreadAssertScope is now always either a subclass, or an alias of
>     PerThreadAssertScopeDebugOnly, so that any function taking a
>     PerThreadAssertScopeDebugOnly* can take the corresponding
>     PerThreadAssertScope* instead.
>
> Change-Id: Id794a330539244c12c082b34354b5671c86c4d35
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5672593
> Reviewed-by: Igor Sheludko <ishell@chromium.org>
> Commit-Queue: Leszek Swirski <leszeks@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#94788}

Change-Id: I961719916dee150e40c069d42fa8c35292738fd5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5675244
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94802}