[compiler] Store isolate_address as ID in ExternalReference
So far `ExternalReference` stored isolate-relative addresses as
absolute addresses. These absolute addresses got converted to offsets
for root-relative loads with pointer arithmetic. However, for this
pointer arithmetic a valid isolate was needed, which is not available
e.g. in WebAssembly compilation. This CL is a first step to allow to
store IDs in the `ExternalReference`. At the moment only the
isolate_address() is handled with the `IsolateDataId`, but other
isolate-relative address should follow eventually.

Bug: 349952800
Change-Id: I31bc38f91cb87a59a66000e8be11fa49eda0f1d7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5645078
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94815}