[heap] Fix DCHECK in clearing of weakrefs
A number of DCHECKs introduced by https://crrev.com/c/5648710, and
in particular the DCHECK enforcing that the value of a weak reference
to be cleared would not be in the RO space, was wrong. It could be
falsified in case the value of the weak reference changed between
marking and clearing. This CL fixes the issue.

Bug: 348181803
Bug: 350256147
Change-Id: Id4582b1992606ed5974e71731fc51dedc9b9ffe5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5676327
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94818}