[isolate-groups] Free code range before reservation is freed
In pointer-compression configurations, the code range uses the
reservation's page allocator; accessing it after the reservation is
freed is a UAF.  This can occur during testing (the PoolTest).

Also, reset the once variable to allow the coderange to be re-created.

Thanks to Milad Fa for the report.

Bug: 42204573
Change-Id: Ia7e89d70e2ade71f8efc8bfe109b575fc00db1b2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5677029
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Andy Wingo <wingo@igalia.com>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94850}