[isel] Avoid putting FrameState constant bitcast input in register
This is a follow-up to https://crrev.com/c/5233669 where this has been
done for word32 constants.

In wasm a
  ref.i31(i32.const 12345)
is translated to:
  1: Int64Constant[24690]
  2: BitcastWordToTaggedSigned(1)
with Turbofan and
  1: Constant[Word64, 24690]
  2: TaggedBitCast[Word64, Tagged, Smi](1)
with Turboshaft.

So, differently to JS, wasm emits Word64 constants for the input of
such constant bitcasts on 64 bit platforms requiring the conversion of
these constants to immediates to be done for Word64 constants as well.

Note that in most cases Turboshaft seems to convert the TaggedBitCast
with constant input to a smi constant prior to instruction selection,
so this fix seems to only apply to corner cases where this optimization
didn't happen before.

Fixed: 350779988
Bug: 42204618, 42202729
Change-Id: Ieb0e3468d6c062e7f920ee5494580bc0438c60fe
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5676696
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94831}