[turboshaft][assembler] Enable short builtin calls
Assembler options should be configured with
`AssemblerOptions::Default(isolate)`, so that if short builtin calls are
enabled in the isolate, the option is passed down from the isolate to
the assembler. Otherwise we generate indirect calls to builtins when
they are in range of direct calls.

Fix it by making users of `TurboShaft::PipelineData()` have to pass down
an `AssemblerOptions` struct, rather than rely on a default argument.

Change-Id: Id4a3a59df1e741a41da00be401ecba80b5b5f317
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5678223
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Pierre Langlois <pierre.langlois@arm.com>
Cr-Commit-Position: refs/heads/main@{#94880}