[cppgc] Fix mutator thread sweeping with deadline
The deadline was recomputed for each space. This was not the intention
of a maximum duration that is passed here. Fix this by only computing
a single deadline that is passed along.

Gardeners: This may shift GC timing by capping sweeping the way we
intended it.

Bug: 333981063
Change-Id: Id631bb89038314f4e782c3f4408f35a07ef5df92
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5679210
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94849}