[objects] Unbox time value in JSDate and optimize TimeClip
Make the date value in JSDate an unboxed float64. This can avoid
allocations of HeapNumbers when setting the value, and a pointer chase
when reading it in optimized code.

Additionally, the TimeClip will, most of the time, leave the value
unchanged. Make it return a boolean for the case where it would have
returned a NaN, so that callers can special case the NaN path (which
usually means throwing or invalidating the JSDate). This also means we
can split up the NaN and non-NaN paths for setting JSDate values, since
we already know in the caller whether we're setting a NaN, which avoids
the NaN check.

Finally, for the specific case of Date.prototype.setTime, allow the
incoming time Number to be returned directly, without a new allocation,
if TimeClip didn't change its value.

Drive-by optimize Object::SameNumberValue to fold together the double
compare and the signbit compare into a single bitwise compare -- this
is ~5% faster on a microbenchmark
(https://quick-bench.com/q/frjnGcEf51jejnsmaG-U0q1t2LQ)

Change-Id: Ic1502d762cc85f2b9daa75d0eb797f327db13685
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5679172
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94856}